Bump dependencies

+++ src/main.rs
@@ -1,5 +1,5 @@
-use clipboard_win::Clipboard;
-use passwords::db;
+use clipboard_win;
+use passwords::db::{Database, FileStatus};
 use rpassword;
 use std::{
     env,
@@ -64,30 +64,31 @@ fn handle_setup() -> Result<String, Box<dyn Error>> {
 
     print_and_flush!("Welcome! ");
 
-    match db::Database::files_exist(&path) {
-        db::FileStatus::None => (),
-        db::FileStatus::Some => {
-            print_and_flush!("It looks like some configuration files are missing. Are you sure you want to overwrite the ones that remain? This will clear the stored data. y/N ");
-            match read_input()?.as_str() {
-                "y" | "Y" => db::Database::delete(&path)?,
-                _ => return Err("Aborting setup".into()),
+    match Database::files_exist(&path) {
+        FileStatus::None => (),
+        FileStatus::Some => {
+            if confirm_with_user("It looks like some configuration files are missing. Are you sure you want to overwrite the ones that remain? This will clear the stored data.")? {
+               Database::delete(&path)?;
+                println!("Previous configuration cleared.");
+            } else {
+                return Err("Aborting setup".into());
             }
         }
-        db::FileStatus::All => {
-            print_and_flush!("It looks like you already have a config ready to go. Are you sure you want to overwrite it? This will clear the stored data. y/N ");
-            match read_input()?.as_str() {
-                "y" | "Y" => db::Database::delete(&path)?,
-                _ => return Err("Aborting setup".into()),
-            };
+        FileStatus::All => {
+            if confirm_with_user("It looks like you already have a config ready to go. Are you sure you want to overwrite it? This will clear the stored data.")? {
+                Database::delete(&path)?;
+                println!("Previous configuration cleared.");
+            } else {
+                return Err("Aborting setup".into());
+            }
         }
     };
 
     let password = loop {
-        print_and_flush!("Please choose a master password: ");
-        let init_password = rpassword::read_password()?;
+        let init_password = rpassword::prompt_password("Please choose a master password: ")?;
 
-        print_and_flush!("Please confirm your master password: ");
-        let confirm_init_password = rpassword::read_password()?;
+        let confirm_init_password =
+            rpassword::prompt_password("Please confirm your master password: ")?;
 
         if init_password == confirm_init_password {
             break init_password;
@@ -96,7 +97,7 @@ fn handle_setup() -> Result<String, Box<dyn Error>> {
         println!("Your passwords don't match. Please try again.");
     };
 
-    db::Database::create_new(&path, &password)?;
+    Database::create_new(&path, &password)?;
 
     Ok("Awesome! You're ready to go.".to_string())
 }
@@ -108,27 +109,27 @@ fn handle_add(maybe_name_to_add: Option<&String>) -> Result<String, Box<dyn Erro
 
     let database = prepare_db_and_password()?;
 
-    match database.get_password(&name_to_add)? {
-        None => {
-            let password_to_add = loop {
-                print_and_flush!("Enter password to add for {}: ", name_to_add);
-                let password_to_add = rpassword::read_password()?;
+    if let Some(_) = database.get_password(name_to_add)? {
+        return Err(format!("You've already saved a password for {name_to_add}").into());
+    };
 
-                print_and_flush!("Confirm password to add for {}: ", name_to_add);
-                let password_confirm = rpassword::read_password()?;
+    let password_to_add = loop {
+        let password_to_add =
+            rpassword::prompt_password(format!("Enter password to add for {name_to_add}: "))?;
 
-                if password_to_add == password_confirm {
-                    break password_to_add;
-                }
+        let password_confirm =
+            rpassword::prompt_password(format!("Confirm password to add for {name_to_add}: "))?;
 
-                println!("The passwords don't match. Please try again.");
-            };
-
-            database.add_password(name_to_add, &password_to_add)?;
-            Ok(format!("Added password for {name_to_add}!"))
+        if password_to_add == password_confirm {
+            break password_to_add;
         }
-        Some(_) => Ok(format!("You've already saved a password for {name_to_add}")),
-    }
+
+        println!("The passwords don't match. Please try again.");
+    };
+
+    database.add_password(name_to_add, &password_to_add)?;
+
+    Ok(format!("Added password for {name_to_add}"))
 }
 
 fn handle_get(maybe_name_to_get: Option<&String>) -> Result<String, Box<dyn Error>> {
@@ -140,36 +141,35 @@ fn handle_get(maybe_name_to_get: Option<&String>) -> Result<String, Box<dyn Erro
 
     match database.get_password(name_to_get)? {
         None => Ok(format!("You haven't saved a password for {name_to_get}")),
-        Some(res) => {
-            Clipboard::new()?.set_string(&res)?;
-            Ok(format!("Copied password for {name_to_get} to clipboard"))
-        }
+        Some(res) => match clipboard_win::set_clipboard_string(&res) {
+            Ok(()) => Ok(format!("Copied password for {name_to_get} to clipboard")),
+            Err(e) => Err(format!("{e}").into()),
+        },
     }
 }
 
 fn handle_all() -> Result<String, Box<dyn Error>> {
     let database = prepare_db_and_password()?;
 
-    print_and_flush!(
-        "Are you sure you want to get all passwords? They will be copied to your clipboard. y/N: "
-    );
+    if !confirm_with_user(
+        "Are you sure you want to get all passwords? They will be copied to your clipboard.",
+    )? {
+        return Ok("Cancelling retrieving all passwords".to_string());
+    }
 
-    match read_input()?.as_str() {
-        "y" | "Y" => {
-            let results = database.get_all_passwords()?;
+    let results = database.get_all_passwords()?;
 
-            if results.len() == 0 {
-                Ok("No passwords found".to_string())
-            } else {
-                let mut joined = String::new();
-                for (name, password) in results {
-                    joined += format!("{name}: {password}\n").as_str();
-                }
-                Clipboard::new()?.set_string(joined.as_str())?;
-                Ok("Copied all passwords to clipboard".to_string())
-            }
-        }
-        _ => Ok("Cancelling retrieving all passwords".to_string()),
+    if results.len() == 0 {
+        return Ok("No passwords found".to_string());
+    }
+
+    let mut joined = String::new();
+    for (name, password) in results {
+        joined += format!("{name}: {password}\n").as_str();
+    }
+    match clipboard_win::set_clipboard_string(joined.as_str()) {
+        Ok(()) => Ok("Copied all passwords to clipboard".to_string()),
+        Err(e) => Err(format!("{e}").into()),
     }
 }
 
@@ -180,24 +180,19 @@ fn handle_remove(maybe_name_to_remove: Option<&String>) -> Result<String, Box<dy
 
     let database = prepare_db_and_password()?;
 
-    match database.get_password(name_to_remove)? {
-        None => Ok(format!("You haven't saved a password for {name_to_remove}")),
-        Some(_) => {
-            print_and_flush!(
-                "Are you sure you want to remove password for {}? y/N: ",
-                name_to_remove
-            );
+    if let None = database.get_password(name_to_remove)? {
+        return Err(format!("You haven't saved a password for {name_to_remove}").into());
+    }
 
-            match read_input()?.as_str() {
-                "y" | "Y" => {
-                    database.remove_password(name_to_remove)?;
-                    Ok(format!(
-                        "Successfully removed password for {name_to_remove}"
-                    ))
-                }
-                _ => Ok("Removal cancelled - no data was affected".to_string()),
-            }
-        }
+    if confirm_with_user(&format!(
+        "Are you sure you want to remove password for {name_to_remove}?"
+    ))? {
+        database.remove_password(name_to_remove)?;
+        Ok(format!(
+            "Successfully removed password for {name_to_remove}"
+        ))
+    } else {
+        Ok("Removal cancelled - no data was affected".to_string())
     }
 }
 
@@ -212,19 +207,18 @@ fn handle_list() -> Result<String, Box<dyn Error>> {
     }
 }
 
-fn prepare_db_and_password() -> Result<db::Database, Box<dyn Error>> {
+fn prepare_db_and_password() -> Result<Database, Box<dyn Error>> {
     let data_dir = get_data_directory()?;
 
-    match db::Database::files_exist(&data_dir) {
-        db::FileStatus::All => (),
-        db::FileStatus::Some => return Err("It looks like some configuration files are missing. Please run passwords setup to get started.".into()),
-        db::FileStatus::None => return Err("It looks like you haven't set up this application yet. Please run passwords setup to get started.".into()),
+    match Database::files_exist(&data_dir) {
+        FileStatus::All => (),
+        FileStatus::Some => return Err("It looks like some configuration files are missing. Please run passwords setup to get started.".into()),
+        FileStatus::None => return Err("It looks like you haven't set up this application yet. Please run passwords setup to get started.".into()),
     };
 
-    print_and_flush!("Enter master password: ");
-    let password = rpassword::read_password()?;
+    let password = rpassword::prompt_password("Enter master password: ")?;
 
-    Ok(db::Database::use_existing(&data_dir, &password)?)
+    Ok(Database::use_existing(&data_dir, &password)?)
 }
 
 fn get_data_directory() -> io::Result<path::PathBuf> {
@@ -232,6 +226,11 @@ fn get_data_directory() -> io::Result<path::PathBuf> {
         .parent()
         .expect("executables are always in a folder")
         .join(".data"))
+}
+
+fn confirm_with_user(message: &str) -> io::Result<bool> {
+    print_and_flush!("{message} y/N: ");
+    Ok(read_input()?.eq_ignore_ascii_case("Y"))
 }
 
 fn read_input() -> io::Result<String> {
